Remove multipdf files. Repair bug in splitpdf. Files prepare to convert into img.

+++ scanpreparation.py
@@ -7,23 +7,20 @@ from PyPDF2 import PdfFileWriter, PdfFileReader
 path = "C:\\Users\\Adam\\Documents\\dane do ml\\pdf"
 COUNT = 1
 
-def renamefiles():
-
+def renamefiles(text):
     os.chdir(path)
-
-
     # Function to increment count
     # to make the files sorted.
     def increment():
         global COUNT
         COUNT = COUNT + 1
 
-    for f in os.listdir():
+    for f in os.listdir(path):
         f_name, f_ext = os.path.splitext(f)
-        f_name = "scans" + str(COUNT)
+        f_name = text + str(COUNT)
         increment()
 
-        new_name = '{} {}'.format(f_name, f_ext)
+        new_name = '{}_zmiana{}'.format(f_name, f_ext)
         os.rename(f, new_name)
 
 
@@ -31,18 +28,22 @@ def renamefiles():
 def splitpdf():
     os.chdir(path)
     for filename in os.listdir(path):
-
-        inputpdf = PdfFileReader(open(filename, "rb"))
+        infile = open(filename, "rb")
+        inputpdf = PdfFileReader(infile, strict=False)
 
         for i in range(inputpdf.numPages):
             output = PdfFileWriter()
             output.addPage(inputpdf.getPage(i))
-            with open("document-{0}-page{1}.pdf".format(filename, i), "wb") as outputStream:
+            with open("s_{0}-page{1}.pdf".format(filename.replace(".pdf",""), i), "wb") as outputStream:
                 output.write(outputStream)
+            outputStream.close()
+        infile.close()  #find this take me a lot of time. It's close open pdf and only after it you can delete old files
+
+
 
 
 def remove_old_files():
     os.chdir(path)
     for filename in os.listdir(path):
-        os.remove(path + "\\" + filename)
-
+        if filename.startswith("scan"):
+            os.remove(path + "\\" + filename)
